requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
dj-database-url==2.1.0
Django==3.2.23
django-axes==6.3.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
django-defender==0.9.7
django-meta==2.4.1
django-redis==5.4.0
django-redis-cache==3.0.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.7.0
packaging==23.2
Pillow==8.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
whitenoise==6.5.0
zipp==3.15.0
